Add copy paste to step context menu (#249)
