+++ packages/react-ui/public/locales/en/translation.json
@@ -110,11 +110,14 @@
   "This step needs to be tested in order to view its data.": "This step needs to be tested in order to view its data.",
   "Go to Trigger": "Go to Trigger",
   "Go to Step": "Go to Step",
+  "Replace": "Replace",
+  "Copy": "Copy",
+  "Duplicate": "Duplicate",
+  "Paste after": "Paste after",
   "True": "True",
   "False": "False",
   "Invalid Move": "Invalid Move",
   "The destination location is inside the same step": "The destination location is inside the same step",
-  "Duplicate": "Duplicate",
   "Incomplete settings": "Incomplete settings",
   "logo": "logo",
   "Step Icon": "Step Icon",
