+++ packages/react-ui/src/app/features/builder/flow-canvas/nodes/step-action-wrapper.tsx
@@ -1,0 +1,12 @@
+import { memo } from 'react';
+
+const StepActionWrapper = memo(
+  ({ children }: { children: React.ReactNode }) => {
+    return (
+      <div className="flex items-center gap-2 cursor-pointer">{children}</div>
+    );
+  },
+);
+
+StepActionWrapper.displayName = 'StepActionWrapper';
+export { StepActionWrapper };
